requirements.txt
aiosqlite
bs4
pandas
pillow
lxml
aiosmtplib
aiofile